Code for the Magic Pumpkin project
The Magic Pumpkin is a simple Halloween project that is powered with a
Pro Trinket 5V that controls a NeoPixel Ring that uses a random #
generator to cycle through patterns. The code is completely portable and
you can power as many lights as you have pins for them since each
neoPixel is its own class.